// atlassian-frontend/packages/monorepo-tooling/dependency-version-analytics/src/commands/populate-historic-data/types.ts
import { DependencyType } from '../../types';

export type PopulateHistoricDataFlags = {
  cwd?: string;
  dev: boolean;
  dryRun: boolean;
  limit?: number;
  interactive: boolean;
};

export type DependencyMap = {
  [name: string]: {
    version: string;
    type: DependencyType;
  };
};

export type AkPackageChange = {
  akDeps: DependencyMap;
  date: string;
};